requirements.txt
numpy==1.21.2
pandas==1.3.4
Pillow==8.4.0
scikit_learn==1.0.2
torch==1.10.0
torchvision==0.11.1
